// src/api/generated/model/firmographicConditionOp.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */

export type FirmographicConditionOp = typeof FirmographicConditionOp[keyof typeof FirmographicConditionOp];


export const FirmographicConditionOp = {
  lt: 'lt',
  gt: 'gt',
  eq: 'eq',
  in: 'in',
  not_in: 'not_in',
  intersects: 'intersects',
} as const;